bug fix implemented a new commands

+++ src/main/kotlin/commands/Clear.kt
@@ -3,4 +3,4 @@ package commands
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class Clear(val guild: Double)
+data class Clear(val guild: Long)
