+++ src/main/kotlin/commands/Pause.kt
@@ -3,4 +3,4 @@ package commands
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class Pause(val guild: Double, val pause: Boolean)
+data class Pause(val guild: Long, val pause: Boolean)
