+++ src/main/kotlin/commands/Seek.kt
@@ -3,4 +3,4 @@ package commands
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class Seek(val guild: Double, val time: Int)
+data class Seek(val guild: Long, val time: Long)
