+++ src/main/kotlin/commands/Shuffle.kt
@@ -3,4 +3,4 @@ package commands
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class Shuffle(val guild: Double)
+data class Shuffle(val guild: Long)
